All Functionality completed

+++ AssessmentProject/Controllers/CourseController.cs
@@ -24,7 +24,7 @@ public class CourseController : Controller
         PaginationViewModel<CourseViewModel> courseVM = _courseService.GetCourseList(search, sortColumn, sortDirection, pageNumber, pageSize);
         return PartialView("_CourseListPartial", courseVM);
     }
-    
+
 
     [HttpGet]
     [Authorize(Roles = "Admin")]
@@ -70,7 +70,16 @@ public class CourseController : Controller
     }
 
     [HttpPost]
-    public async Task<IActionResult> DeleteCourse(int id){
+    public async Task<IActionResult> DeleteCourse(int id)
+    {
+
+        var isAnyEnrolled = _courseService.isAnyEnrolled(id);
+
+        if(isAnyEnrolled){
+             return Json(new { success = false, message = "Course can't be Deleted As Students are enrolled" });
+        }
+
+
         var isDeleted = await _courseService.DeleteCourse(id);
         if (isDeleted)
         {
@@ -84,26 +93,63 @@ public class CourseController : Controller
     public IActionResult GetCourseListStudent(string search, string sortColumn, string sortDirection, int pageNumber, int pageSize)
     {
         var studentId = int.Parse(Request.Cookies["UserId"]);
-        PaginationViewModel<CourseViewModel> courseVM = _courseService.GetCourseListStudent(studentId,search, sortColumn, sortDirection, pageNumber, pageSize);
+        PaginationViewModel<CourseViewModel> courseVM = _courseService.GetCourseListStudent(studentId, search, sortColumn, sortDirection, pageNumber, pageSize);
         return PartialView("_CourseListPartial", courseVM);
     }
 
 
-    
+
     [HttpPost]
-    public async Task<IActionResult> EnrollCourse(int courseId){
+    public async Task<IActionResult> EnrollCourse(int courseId)
+    {
         var studentId = int.Parse(Request.Cookies["UserId"]);
+        
+        var IsAlreadyCompleted = _courseService.IsAlreadyCompleted(courseId, studentId);
+
+        if (IsAlreadyCompleted)
+        {
+            return Json(new { success = false, message = "Already Completed!!!" });
+        }
+
         var isAlreadyEnrolled = _courseService.IsAlreadyEnrolled(courseId, studentId);
 
-        if(isAlreadyEnrolled){
+        if (isAlreadyEnrolled)
+        {
             return Json(new { success = false, message = "Already Enrolled!!!" });
         }
 
-        var isDeleted = await _courseService.EnrollCourse(courseId, studentId);
-        if (isDeleted)
+        var isEnrolled = await _courseService.EnrollCourse(courseId, studentId);
+        if (isEnrolled)
         {
             return Json(new { success = true, message = "Course Enrolled successfully" });
         }
         return Json(new { success = false, message = "Course Not Enrolled" });
     }
+
+    [HttpPost]
+    public async Task<IActionResult> WithdrawCourse(int courseId)
+    {
+        var studentId = int.Parse(Request.Cookies["UserId"]);
+
+        var isWithdrawn = await _courseService.WithdrawCourse(courseId, studentId);
+        if (isWithdrawn)
+        {
+            return Json(new { success = true, message = "Course Withdrawn successfully" });
+        }
+        return Json(new { success = false, message = "Course Not Withdrawn" });
+    }
+
+    [HttpPost]
+    public async Task<IActionResult> CompleteCourse(int courseId)
+    {
+        var studentId = int.Parse(Request.Cookies["UserId"]);
+
+        var isCompleted = await _courseService.CompleteCourse(courseId, studentId);
+        if (isCompleted)
+        {
+            return Json(new { success = true, message = "Course Completed successfully" });
+        }
+        return Json(new { success = false, message = "Course Not Completed" });
+    }
+
 }
